ai-service/requirements.txt

fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-dotenv==1.0.0
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
numpy==1.26.2
pillow==10.1.0
opencv-python==4.9.0.80
requests==2.31.0
scikit-learn==1.4.0
tensorflow==2.15.0